Began porting js physics to java  (NOT FINAL)

+++ Server/src/Uinput.java
@@ -1,8 +1,8 @@
 public class Uinput {
 	public int type;			// 0 for a mouse move, 1 for a mouse click
-	public float xdir, ydir; 
+	public double xdir, ydir;
 	public int pid;
-	public Uinput(int Type, int pId, float xDir, float yDir)	{
+	public Uinput(int Type, int pId, double xDir, double yDir)	{
 		pid  = pId;
 		xdir = xDir;
 		ydir = yDir;
